changed types

+++ arduino/arduino_main/arduino_main.ino
@@ -19,11 +19,11 @@ const char topic_speed[] = "vehicle-crash-detection/sensors/gps/speed";
 
 
 // sample rate
-const int SAMPLE_RATE = 10;
+const uint8_t  SAMPLE_RATE = 10;
 
 // impact sensor
-const int impact_pin = 3;
-int impact_value = 0;
+const uint8_t impact_pin = 3;
+uint8_t impact_value = 0;
 
 
 // starting program
@@ -59,7 +59,7 @@ void loop() {
 
 // initializing wifi connection
 void connect_wifi() {
-  int status = WL_IDLE_STATUS;
+  uint8_t status = WL_IDLE_STATUS;
   while (status != WL_CONNECTED) {
     Serial.print("Attempting to connect to network: ");
     Serial.println(ssid);
